+++ app/renderer/src/main/src/pages/layout/mainOperatorContent/MainOperatorContent.module.scss
@@ -1,4 +1,5 @@
 @use "../../../assets/global.scss" as global;
+
 $colors: global.$colors;
 $keyColor: map-keys($colors);
 @for $c from 1 through length($keyColor) {
@@ -54,7 +55,6 @@ $keyColor: map-keys($colors);
 .tab-menu {
     display: flex;
     flex-direction: column;
-    margin: 20px;
     font-size: 12px;
     overflow: hidden;
     .tab-menu-first {
